complete 4.21 adding authenticated user to posted blogs deletion can now only occur if user is the same user that created blog

+++ controllers/blog.js
@@ -10,8 +10,8 @@ blogRoute.get("/", async (_request, response) => {
 });
 
 blogRoute.post("/", async (request, response) => {
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-  if (!decodedToken.id) {
+  const { id: userId } = jwt.verify(request.token, process.env.SECRET);
+  if (!userId) {
     return response.status(401).json({ error: "token invalid" });
   }
   const { title, url, author, likes, _id } = request.body;
@@ -24,8 +24,7 @@ blogRoute.post("/", async (request, response) => {
     return response.status(400).json({ error: "URL is missing from request" });
   }
 
-  const users = await User.find({});
-  const user = users[0];
+  const user = await User.findById(userId);
   const blogToCreate = {
     title,
     url,
@@ -59,8 +58,19 @@ blogRoute.put("/:id", async (request, response) => {
 });
 
 blogRoute.delete("/:id", async (request, response) => {
+  const { id: userId } = jwt.verify(request.token, process.env.SECRET);
+  if (!userId) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.user.toString() !== userId.toString()) {
+    return response.status(401).json({ error: "unauthorized" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  return response.status(204).end();
 });
 
 module.exports = blogRoute;
